set voiceCapture as dependency for SpeachRecognizer sperate classes file from protocols files
