+++ Arabic Voice Recognition/VoiceCapture.swift
@@ -14,26 +14,3 @@ protocol VoiceCapture {
     func stop()
     
 }
-class AVFoundationVoiceCapture: VoiceCapture {
-    private let audioEngine = AVAudioEngine()
-    
-    func startCapture(completion: @escaping (AVAudioPCMBuffer)->()) throws {
-        let audioSession = AVAudioSession.sharedInstance()
-        try audioSession.setCategory(.record, mode: .measurement, options: .duckOthers)
-        try audioSession.setActive(true, options: .notifyOthersOnDeactivation)
-        let inputNode = audioEngine.inputNode
-        let recordingFormat = inputNode.outputFormat(forBus: 0)
-        inputNode.installTap(onBus: 0, bufferSize: 1024, format: recordingFormat) { (buffer: AVAudioPCMBuffer, when: AVAudioTime) in
-            completion(buffer)
-        }
-        
-        audioEngine.prepare()
-        try audioEngine.start()
-    }
-    
-    func stop() {
-        
-    }
-    
-    
-}
